--- src/Modal/utils/helpers.ts
@@ -1,1 +1,0 @@
-export type ModalSizes = 'sm' | 'md' | 'lg';
